qdts_orchestrator/pyproject.toml
[tool.poetry]
name = "qdts_orchestrator"
version = "0.1.8"
description = ""
authors = ["rmcelaya <rmcelaya@protonmail.com>"]

[tool.poetry.dependencies]
pyyaml="*"
ansible="*"
ansible_runner="*"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
qdts_run = "qdts_orchestrator.run:run"
qdts_stop = "qdts_orchestrator.stop:stop"
qdts_install="qdts_orchestrator.install:install"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
